Rework the layout of the campaigns #7
Use a dynamic height of the cell and make it dependent on the cell's contents.
The image should take the full available width of the screen and dynamically adjust its height with an aspect ratio of 4:3.
The text should go below the image. There should not be any space between the title text and the image, because the image already contains whitespace.
The text should have a padding of 8 points to the left and to the right.
There should be 8 points spacing between the title text and the description text.
The title text should break into two lines at most. Every text which is longer than two lines should be cut after the second line.
The description text should not have any length restrictions.